MetricsConfiguration is just an enum-typed dict
We're iterating over the keys often enough that we might as well just make it official instead of calling getattr() so much.
